refactor(tasks/boards): modularize enhanced-boards-view into focused components
🔧 BREAKING DOWN MONOLITHIC FILE (3497 → 1722 lines)

✨ New Components Created:
- components/GanttChart.tsx (1027 lines) - Timeline visualization with pagination & search
- components/TaskWorkflowAnalytics.tsx (208 lines) - Workflow efficiency metrics
- components/TaskCreationAnalytics.tsx (214 lines) - Creation trends & analytics
- components/StatusDistribution.tsx (117 lines) - Task status distribution charts
- components/TaskGroup.tsx (230 lines) - Collapsible task sections with priority badges

🛠️ Utilities & Hooks:
- utils/taskHelpers.ts (177 lines) - Shared utility functions for task operations
- hooks/useTaskAnalytics.ts (247 lines) - Analytics calculation hooks

🐛 Critical React Hooks Fixes:
- Fixed hook order violations in GanttChart component
- Moved useMemo from JSX to top-level (Rules of Hooks compliance)
- Resolved 'Rendered more hooks than during the previous render' error

🧹 Code Quality Improvements:
- Removed all unused imports and dependencies
- Clean separation of concerns between components
- Maintained full functionality while improving maintainability
- Each component now has focused, single responsibility

📊 Benefits:
- ✅ 50% reduction in main file size (3497 → 1722 lines)
- ✅ Improved testability with isolated components
- ✅ Enhanced maintainability and code readability
- ✅ Resolved React Hooks runtime errors
- ✅ Better developer experience for future modifications

This refactoring transforms a large, difficult-to-maintain monolithic component into a well-structured, modular architecture while preserving all existing functionality and fixing critical React compliance issues.
