+++ apps/web/src/app/[locale]/(dashboard)/[wsId]/tasks/boards/enhanced-boards-view.tsx
@@ -7,14 +7,17 @@ import {
   TaskBoard,
   TaskList,
 } from '@tuturuuu/types/primitives/TaskBoard';
+// Import new components
+import { GanttChart } from './components/GanttChart';
+import { StatusDistribution } from './components/StatusDistribution';
+import { TaskWorkflowAnalytics } from './components/TaskWorkflowAnalytics';
+import { TaskCreationAnalytics } from './components/TaskCreationAnalytics';
+import { TaskGroup } from './components/TaskGroup';
+
 import { Badge } from '@tuturuuu/ui/badge';
 import { Button } from '@tuturuuu/ui/button';
 import { Card } from '@tuturuuu/ui/card';
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from '@tuturuuu/ui/collapsible';
+
 import {
   Dialog,
   DialogContent,
@@ -28,11 +31,7 @@ import {
   BarChart3,
   Calendar,
   CheckCircle2,
-  ChevronDown,
-  ChevronLeft,
-  ChevronRight,
   Clock,
-  ExternalLink,
   // Users,
   Eye,
   Filter,
@@ -56,7 +55,7 @@ import {
 } from '@tuturuuu/ui/select';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@tuturuuu/ui/tabs';
 import { cn } from '@tuturuuu/utils/format';
-import { useEffect, useMemo, useRef, useState } from 'react';
+import { useMemo, useState } from 'react';
 
 interface AnalyticsFilters {
   timeView: 'week' | 'month' | 'year';
@@ -71,1057 +70,6 @@ const CARD_LAYOUT_OPTIONS = [
 ] as const;
 
 type CardLayout = typeof CARD_LAYOUT_OPTIONS[number]['value'];
-
-/**
- * Calculate the number of days a task is overdue
- * @param dueDate The task's due date
- * @returns Number of days the task is overdue
- */
-function calculateOverdueDays(dueDate: string | Date): number {
-  const due = new Date(dueDate);
-  return Math.ceil((new Date().getTime() - due.getTime()) / (1000 * 60 * 60 * 24));
-}
-
-// Gantt Chart Component
-function GanttChart({
-  allTasks,
-  filters,
-}: {
-  allTasks: any[];
-  filters: AnalyticsFilters;
-}) {
-  const [currentPage, setCurrentPage] = useState(1);
-  const [pageSize, setPageSize] = useState(5);
-  const [searchQuery, setSearchQuery] = useState('');
-  const [clickCardVisible, setClickCardVisible] = useState(false);
-  const [clickCardPosition, setClickCardPosition] = useState({ x: 0, y: 0 });
-  const [clickedTask, setClickedTask] = useState<any | null>(null);
-  const clickCardRef = useRef<HTMLDivElement>(null);
-
-  // Filter tasks based on selected board, status, and search
-  const filteredTasks = useMemo(() => {
-    let tasks = allTasks;
-
-    // Filter by board
-    if (filters.selectedBoard) {
-      tasks = tasks.filter((task) => task.boardId === filters.selectedBoard);
-    }
-
-    // Filter by status
-    if (filters.statusFilter !== 'all') {
-      tasks = tasks.filter((task) => {
-        const taskStatus = task.listStatus || 'not_started';
-        return taskStatus === filters.statusFilter;
-      });
-    }
-
-    // Filter by search query
-    if (searchQuery.trim()) {
-      const query = searchQuery.toLowerCase();
-      tasks = tasks.filter(
-        (task) =>
-          task.name?.toLowerCase().includes(query) ||
-          task.description?.toLowerCase().includes(query) ||
-          task.boardName?.toLowerCase().includes(query)
-      );
-    }
-
-    return tasks;
-  }, [allTasks, filters.selectedBoard, filters.statusFilter, searchQuery]);
-
-  // Get time range based on filter
-  const getTimeRange = () => {
-    const now = new Date();
-    switch (filters.timeView) {
-      case 'week': {
-        const weekStart = new Date(now);
-        weekStart.setDate(now.getDate() - now.getDay());
-        const weekEnd = new Date(weekStart);
-        weekEnd.setDate(weekStart.getDate() + 6);
-        return { start: weekStart, end: weekEnd };
-      }
-      case 'month': {
-        const monthStart = new Date(now.getFullYear(), now.getMonth(), 1);
-        const monthEnd = new Date(now.getFullYear(), now.getMonth() + 1, 0);
-        return { start: monthStart, end: monthEnd };
-      }
-      case 'year': {
-        const yearStart = new Date(now.getFullYear(), 0, 1);
-        const yearEnd = new Date(now.getFullYear(), 11, 31);
-        return { start: yearStart, end: yearEnd };
-      }
-    }
-  };
-
-  const timeRange = getTimeRange();
-  const totalDays = Math.ceil(
-    (timeRange.end.getTime() - timeRange.start.getTime()) /
-      (1000 * 60 * 60 * 24)
-  );
-
-  // Calculate productivity stats
-  const productivityStats = useMemo(() => {
-    const completed = filteredTasks.filter(
-      (task) => task.listStatus === 'done' || task.listStatus === 'closed'
-    ).length;
-    const total = filteredTasks.length;
-    const completionRate = total > 0 ? (completed / total) * 100 : 0;
-
-    const overdue = filteredTasks.filter(
-      (task) =>
-        task.end_date &&
-        new Date(task.end_date) < new Date() &&
-        task.listStatus !== 'done' &&
-        task.listStatus !== 'closed'
-    ).length;
-
-    const onTime = completed - overdue;
-    const onTimeRate = completed > 0 ? (onTime / completed) * 100 : 0;
-
-    return {
-      completionRate: completionRate.toFixed(1),
-      onTimeRate: onTimeRate.toFixed(1),
-      totalTasks: total,
-      completedTasks: completed,
-      overdueTasks: overdue,
-    };
-  }, [filteredTasks]);
-
-  // Process all tasks for Gantt display (before pagination)
-  const allGanttTasks = useMemo(() => {
-    return filteredTasks
-      .filter((task) => task.created_at) // Only tasks with creation date
-      .map((task) => {
-        const createdDate = new Date(task.created_at);
-        const endDate =
-          task.listStatus === 'done' || task.listStatus === 'closed'
-            ? task.updated_at
-              ? new Date(task.updated_at)
-              : createdDate
-            : task.end_date
-              ? new Date(task.end_date)
-              : new Date();
-
-        // Calculate position and width
-        const startOffset = Math.max(
-          0,
-          (createdDate.getTime() - timeRange.start.getTime()) /
-            (1000 * 60 * 60 * 24)
-        );
-        const duration = Math.max(
-          1,
-          (endDate.getTime() - createdDate.getTime()) / (1000 * 60 * 60 * 24)
-        );
-
-        return {
-          ...task,
-          createdDate,
-          endDate,
-          startOffset: (startOffset / totalDays) * 100,
-          width: Math.min(
-            (duration / totalDays) * 100,
-            100 - (startOffset / totalDays) * 100
-          ),
-          status: task.listStatus || 'not_started',
-        };
-      })
-      .filter((task) => task.startOffset < 100) // Only show tasks within time range
-      .sort((a, b) => a.createdDate.getTime() - b.createdDate.getTime());
-  }, [filteredTasks, timeRange, totalDays]);
-
-  // Calculate pagination
-  const totalTasks = allGanttTasks.length;
-  const totalPages = Math.ceil(totalTasks / pageSize);
-  const startIndex = (currentPage - 1) * pageSize;
-  const endIndex = startIndex + pageSize;
-
-  // Get current page tasks
-  const ganttTasks = useMemo(() => {
-    return allGanttTasks.slice(startIndex, endIndex);
-  }, [allGanttTasks, startIndex, endIndex]);
-
-  // Reset to first page when filters change
-  useEffect(() => {
-    setCurrentPage(1);
-  }, [filters.selectedBoard, filters.statusFilter, searchQuery]);
-
-  const getStatusColor = (status: string) => {
-    switch (status) {
-      case 'done':
-      case 'closed':
-        return 'bg-green-500';
-      case 'active':
-        return 'bg-blue-500';
-      case 'not_started':
-      default:
-        return 'bg-gray-400';
-    }
-  };
-
-  // Generate time markers
-  const timeMarkers = useMemo(() => {
-    const markers = [];
-    const markerCount =
-      filters.timeView === 'week' ? 7 : filters.timeView === 'month' ? 4 : 12;
-
-    for (let i = 0; i <= markerCount; i++) {
-      const position = (i / markerCount) * 100;
-      const date = new Date(timeRange.start);
-
-      if (filters.timeView === 'week') {
-        date.setDate(date.getDate() + i);
-        markers.push({
-          position,
-          label: date.toLocaleDateString('en', { weekday: 'short' }),
-        });
-      } else if (filters.timeView === 'month') {
-        date.setDate(date.getDate() + i * 7);
-        markers.push({ position, label: `Week ${i + 1}` });
-      } else {
-        date.setMonth(date.getMonth() + i);
-        markers.push({
-          position,
-          label: date.toLocaleDateString('en', { month: 'short' }),
-        });
-      }
-    }
-    return markers;
-  }, [filters.timeView, timeRange]);
-
-  // Add the click handlers to the GanttChart component
-  const handleTaskClick = (e: React.MouseEvent, task: any) => {
-    e.stopPropagation();
-
-    const mouseX = e.clientX;
-    const mouseY = e.clientY;
-
-    setClickedTask(task);
-    setClickCardVisible(true);
-
-    // Position near the click
-    let newX = mouseX + 15;
-    let newY = mouseY + 15;
-
-    // Ensure it doesn't go off screen
-    if (newX + 250 > window.innerWidth) {
-      newX = mouseX - 265;
-    }
-    if (newY + 150 > window.innerHeight) {
-      newY = mouseY - 165;
-    }
-
-    setClickCardPosition({ x: newX, y: newY });
-  };
-
-  const handleCloseClick = () => {
-    setClickCardVisible(false);
-    setClickedTask(null);
-  };
-
-  // Local utility function to get completion date
-  const getTaskCompletionDate = (task: any): Date | null => {
-    const taskData = task as any;
-    const possibleFields = ['updated_at', 'completed_at', 'closed_at', 'finished_at', 'done_at'];
-    
-    // First try to get explicit completion dates
-    for (const field of possibleFields) {
-      const dateStr = taskData[field];
-      if (dateStr) {
-        const date = new Date(dateStr);
-        if (!isNaN(date.getTime())) {
-          return date;
-        }
-      }
-    }
-    
-    // Fallback: if task is completed but no completion date, use updated_at or created_at as estimate
-    if (task.listStatus === 'done' || task.listStatus === 'closed' || task.archived) {
-      // Try updated_at first (might indicate when status was changed)
-      if (taskData.updated_at) {
-        const date = new Date(taskData.updated_at);
-        if (!isNaN(date.getTime())) {
-          return date;
-        }
-      }
-      
-      // Last resort: use creation date (not ideal but better than nothing)
-      if (task.created_at) {
-        const date = new Date(task.created_at);
-        if (!isNaN(date.getTime())) {
-          return date;
-        }
-      }
-    }
-    
-    return null;
-  };
-
-  // Add click outside to close
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        clickCardRef.current &&
-        !clickCardRef.current.contains(event.target as Node)
-      ) {
-        handleCloseClick();
-      }
-    };
-
-    const handleScroll = () => {
-      handleCloseClick();
-    };
-
-    const handleEscape = (event: KeyboardEvent) => {
-      if (event.key === 'Escape') {
-        handleCloseClick();
-      }
-    };
-
-    if (clickCardVisible) {
-      document.addEventListener('mousedown', handleClickOutside);
-      document.addEventListener('scroll', handleScroll, true); // Use capture to catch all scroll events
-      document.addEventListener('keydown', handleEscape);
-      window.addEventListener('scroll', handleScroll);
-      // Also listen for scrolling on the gantt container
-      const ganttContainer = document.querySelector('.custom-scrollbar');
-      if (ganttContainer) {
-        ganttContainer.addEventListener('scroll', handleScroll);
-      }
-    }
-
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-      document.removeEventListener('scroll', handleScroll, true);
-      document.removeEventListener('keydown', handleEscape);
-      window.removeEventListener('scroll', handleScroll);
-      const ganttContainer = document.querySelector('.custom-scrollbar');
-      if (ganttContainer) {
-        ganttContainer.removeEventListener('scroll', handleScroll);
-      }
-    };
-  }, [clickCardVisible]);
-
-  return (
-    <>
-      <Card className="p-6">
-        <div className="mb-6 flex items-center justify-between">
-          <div>
-            <h4 className="font-medium">Task Gantt Timeline</h4>
-            <p className="text-sm text-muted-foreground">
-              Visual timeline showing task lifecycle from creation to completion
-            </p>
-          </div>
-
-          {/* Productivity Stats */}
-          <div className="flex items-center gap-4 text-sm">
-            <div className="text-center">
-              <div className="font-semibold text-green-600">
-                {productivityStats.completionRate}%
-              </div>
-              <div className="text-muted-foreground">Completion</div>
-            </div>
-            <div className="text-center">
-              <div className="font-semibold text-blue-600">
-                {productivityStats.onTimeRate}%
-              </div>
-              <div className="text-muted-foreground">On-Time</div>
-            </div>
-            <div className="text-center">
-              <div className="font-semibold text-purple-600">
-                {productivityStats.totalTasks}
-              </div>
-              <div className="text-muted-foreground">Tasks</div>
-            </div>
-          </div>
-        </div>
-
-        {/* Combined Search, Pagination and Task Count Controls */}
-        <div className="mb-4 flex flex-col items-center justify-between gap-4 md:flex-row">
-          {/* Search Input */}
-          <div className="relative w-full md:max-w-sm">
-            <input
-              type="text"
-              placeholder="Search tasks..."
-              value={searchQuery}
-              onChange={(e) => setSearchQuery(e.target.value)}
-              className="w-full rounded-md border border-gray-200 bg-background px-3 py-2 text-sm focus:ring-2 focus:ring-primary/20 focus:outline-none dark:border-gray-700"
-            />
-            {searchQuery && (
-              <button
-                onClick={() => setSearchQuery('')}
-                className="absolute top-1/2 right-2 -translate-y-1/2 text-muted-foreground hover:text-foreground"
-              >
-                <X className="h-4 w-4" />
-              </button>
-            )}
-          </div>
-
-          {/* Pagination and Task Count */}
-          <div className="flex flex-wrap items-center justify-center gap-4 md:justify-end">
-            <Select
-              value={pageSize.toString()}
-              onValueChange={(value) => {
-                setPageSize(Number(value));
-                setCurrentPage(1);
-              }}
-            >
-              <SelectTrigger className="w-24">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="5">5</SelectItem>
-                <SelectItem value="10">10</SelectItem>
-                <SelectItem value="25">25</SelectItem>
-                <SelectItem value="50">50</SelectItem>
-              </SelectContent>
-            </Select>
-
-            {totalTasks > 0 && (
-              <div className="flex items-center gap-2 text-sm text-muted-foreground">
-                <span>
-                  Showing {startIndex + 1}-{Math.min(endIndex, totalTasks)} of{' '}
-                  {totalTasks} tasks
-                </span>
-
-                {totalPages > 1 && (
-                  <div className="ml-4 flex items-center gap-1">
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() =>
-                        setCurrentPage((prev) => Math.max(1, prev - 1))
-                      }
-                      disabled={currentPage === 1}
-                      className="h-8 w-8 p-0"
-                    >
-                      <ChevronLeft className="h-4 w-4" />
-                    </Button>
-
-                    <span className="px-2 text-sm">
-                      {currentPage} of {totalPages}
-                    </span>
-
-                    <Button
-                      variant="outline"
-                      size="sm"
-                      onClick={() =>
-                        setCurrentPage((prev) => Math.min(totalPages, prev + 1))
-                      }
-                      disabled={currentPage === totalPages}
-                      className="h-8 w-8 p-0"
-                    >
-                      <ChevronRight className="h-4 w-4" />
-                    </Button>
-                  </div>
-                )}
-              </div>
-            )}
-          </div>
-        </div>
-
-        {/* Time Scale - Horizontally Scrollable for Year View */}
-        <div className="relative mb-4 h-10">
-          <div
-            className={cn(
-              'relative border-b pb-3',
-              filters.timeView === 'year'
-                ? 'overflow-x-auto overflow-y-hidden'
-                : 'overflow-hidden'
-            )}
-          >
-            <div
-              className={cn(
-                'relative flex',
-                filters.timeView === 'year'
-                  ? 'h-6 min-w-[1000px]'
-                  : 'h-6 w-full'
-              )}
-            >
-              {timeMarkers.map((marker, index) => (
-                <div
-                  key={index}
-                  className={cn(
-                    'flex items-center justify-center text-xs whitespace-nowrap text-muted-foreground',
-                    filters.timeView === 'year'
-                      ? 'flex-1 px-2 text-center'
-                      : 'absolute -translate-x-1/2 transform'
-                  )}
-                  style={
-                    filters.timeView === 'year'
-                      ? {}
-                      : {
-                          left: `${Math.min(Math.max(marker.position, 8), 92)}%`,
-                          top: '0px',
-                        }
-                  }
-                >
-                  {marker.label}
-                </div>
-              ))}
-            </div>
-          </div>
-        </div>
-
-        {/* Gantt Chart - Fixed Height Container */}
-        <div
-          className="relative rounded-lg border bg-background"
-          style={{ height: '320px' }} // Fixed height instead of expanding
-        >
-          <style
-            dangerouslySetInnerHTML={{
-              __html: `
-            .custom-scrollbar::-webkit-scrollbar {
-              width: 8px;
-            }
-            .custom-scrollbar::-webkit-scrollbar-track {
-              background: transparent;
-            }
-            .custom-scrollbar::-webkit-scrollbar-thumb {
-              background: #3b82f6;
-              border-radius: 4px;
-            }
-            .custom-scrollbar::-webkit-scrollbar-thumb:hover {
-              background: #2563eb;
-            }
-            .hover-card {
-              position: fixed;
-              z-index: 9999;
-              pointer-events: none;
-              opacity: 0;
-              transform: scale(0.95);
-              transition: all 0.2s ease-out;
-              will-change: transform, opacity, left, top;
-            }
-            .group:hover .hover-card {
-              opacity: 1;
-              transform: scale(1);
-              pointer-events: auto;
-            }
-            .hover-card::before {
-              content: '';
-              position: absolute;
-              width: 0;
-              height: 0;
-              border-style: solid;
-              border-width: 5px 5px 5px 0;
-              border-color: transparent #ffffff transparent transparent;
-              left: -5px;
-              top: 12px;
-              filter: drop-shadow(-1px 0 1px rgba(0,0,0,0.1));
-            }
-            .dark .hover-card::before {
-              border-color: transparent #111827 transparent transparent;
-            }
-          `,
-            }}
-          />
-
-          {/* Scrollable Content Area */}
-          <div
-            className="custom-scrollbar h-full overflow-y-auto p-4 pr-2"
-            style={{
-              scrollbarWidth: 'thin',
-              scrollbarColor: '#3b82f6 transparent',
-            }}
-          >
-            <div className="space-y-1">
-              {ganttTasks.length === 0 ? (
-                <div className="py-8 text-center text-muted-foreground">
-                  No tasks found for the selected time period
-                </div>
-              ) : (
-                ganttTasks.map((task) => (
-                  <div
-                    key={task.id}
-                    className="group relative flex items-center gap-3 rounded-lg px-2 py-2 transition-colors hover:bg-gray-50 dark:hover:bg-gray-800/50"
-                  >
-                    {/* Task Info */}
-                    <div className="w-52 min-w-0">
-                      {/* Task Name with line clamp */}
-                      <div
-                        className="line-clamp-1 text-sm font-medium text-gray-900 transition-all duration-200 group-hover:line-clamp-none dark:text-gray-100"
-                        title={task.name}
-                      >
-                        {task.name}
-                      </div>
-                    </div>
-
-                    {/* Timeline Bar */}
-                    <div className="relative h-6 flex-1 rounded bg-gray-100 dark:bg-gray-800">
-                      <div
-                        className={cn(
-                          'absolute h-full cursor-pointer rounded transition-all hover:opacity-90',
-                          getStatusColor(task.status)
-                        )}
-                        style={{
-                          left: `${task.startOffset}%`,
-                          width: `${task.width}%`,
-                        }}
-                      >
-                        {/* Status Indicators */}
-                        <div className="pointer-events-none absolute inset-y-0 flex w-full items-center justify-between">
-                          {/* Not Started (Gray) */}
-                          <div
-                            className="h-2 w-2 rounded-full bg-gray-400 opacity-70"
-                            style={{ left: '0%' }}
-                          />
-
-                          {/* In Progress (Blue) */}
-                          {task.status === 'active' && (
-                            <div
-                              className="h-2 w-2 rounded-full bg-blue-500 opacity-70"
-                              style={{
-                                left: '50%',
-                                transform: 'translateX(-50%)',
-                              }}
-                            />
-                          )}
-
-                          {/* Completed (Green) */}
-                          {(task.status === 'done' ||
-                            task.status === 'closed') && (
-                            <div
-                              className="h-2 w-2 rounded-full bg-green-500 opacity-70"
-                              style={{ right: '0%' }}
-                            />
-                          )}
-                        </div>
-
-                        <div className="flex h-full items-center justify-center text-xs font-medium text-white">
-                          {(task.status === 'done' ||
-                            task.status === 'closed') &&
-                          filters.timeView !== 'year' &&
-                          task.width > 15
-                            ? '✓'
-                            : ''}
-                        </div>
-
-                        {/* Status transition markers and timeline */}
-                        {task.updated_at &&
-                          task.updated_at !== task.created_at && (
-                            <>
-                              {/* Status change marker */}
-                              <div
-                                className="absolute top-0 h-full w-0.5 bg-yellow-400 opacity-60"
-                                style={{ left: '50%' }}
-                                title={`Status changed: ${new Date(task.updated_at).toLocaleDateString()}`}
-                              />
-
-                              {/* Progress indicator for active tasks */}
-                              {task.status === 'active' && (
-                                <div className="absolute inset-0 animate-pulse bg-gradient-to-r from-transparent via-white/20 to-transparent" />
-                              )}
-                            </>
-                          )}
-
-                        {/* Task lifecycle phases - Individual phase hover targets */}
-                        <div className="absolute inset-0 flex">
-                          {/* Creation phase (first 25% of timeline) */}
-                          <div
-                            className="cursor-help bg-gray-300 opacity-30 transition-opacity hover:opacity-50 dark:bg-gray-600"
-                            style={{ width: '25%' }}
-                            title={`📅 Created: ${task.createdDate.toLocaleDateString()} at ${task.createdDate.toLocaleTimeString()}`}
-                          />
-
-                          {/* Active development phase (middle 50%) */}
-                          {task.status === 'active' && (
-                            <div
-                              className="cursor-help bg-blue-400 opacity-40 transition-opacity hover:opacity-60"
-                              style={{ width: '50%' }}
-                              title={`🔄 In Progress: Started ${task.createdDate.toLocaleDateString()}${task.updated_at ? ` • Last updated: ${new Date(task.updated_at).toLocaleDateString()}` : ''}`}
-                            />
-                          )}
-
-                          {/* Completion phase (last 25% if completed) */}
-                          {(task.status === 'done' ||
-                            task.status === 'closed') && (
-                            <div
-                              className="ml-auto cursor-help bg-green-400 opacity-40 transition-opacity hover:opacity-60"
-                              style={{ width: '25%' }}
-                              title={`✅ ${task.status === 'done' ? 'Completed' : 'Closed'}: ${task.updated_at ? new Date(task.updated_at).toLocaleDateString() + ' at ' + new Date(task.updated_at).toLocaleTimeString() : 'Date unknown'}${task.end_date ? ` • Due was: ${new Date(task.end_date).toLocaleDateString()}` : ''}`}
-                            />
-                          )}
-
-                          {/* Main timeline click area */}
-                          <div
-                            className="absolute inset-0 cursor-pointer"
-                            onClick={(e) => handleTaskClick(e, task)}
-                          />
-                        </div>
-                      </div>
-                    </div>
-
-                    {/* Status Badge */}
-                    <div className="w-24 text-xs">
-                      <Badge
-                        variant="outline"
-                        className={cn(
-                          'border text-xs font-medium',
-                          task.status === 'done' &&
-                            'border-green-200 bg-green-50 text-green-700 dark:border-green-800 dark:bg-green-900/20 dark:text-green-400',
-                          task.status === 'closed' &&
-                            'border-purple-200 bg-purple-50 text-purple-700 dark:border-purple-800 dark:bg-purple-900/20 dark:text-purple-400',
-                          task.status === 'active' &&
-                            'border-blue-200 bg-blue-50 text-blue-700 dark:border-blue-800 dark:bg-blue-900/20 dark:text-blue-400',
-                          task.status === 'not_started' &&
-                            'border-gray-200 bg-gray-50 text-gray-700 dark:border-gray-800 dark:bg-gray-900/20 dark:text-gray-400'
-                        )}
-                      >
-                        {task.status === 'done' ? (
-                          <span className="flex items-center gap-1">
-                            <span className="h-1.5 w-1.5 rounded-full bg-green-500"></span>
-                            Done
-                          </span>
-                        ) : task.status === 'closed' ? (
-                          <span className="flex items-center gap-1">
-                            <span className="h-1.5 w-1.5 rounded-full bg-purple-500"></span>
-                            Closed
-                          </span>
-                        ) : task.status === 'active' ? (
-                          <span className="flex items-center gap-1">
-                            <span className="h-1.5 w-1.5 animate-pulse rounded-full bg-blue-500"></span>
-                            Active
-                          </span>
-                        ) : (
-                          <span className="flex items-center gap-1">
-                            <span className="h-1.5 w-1.5 rounded-full bg-gray-400"></span>
-                            Pending
-                          </span>
-                        )}
-                      </Badge>
-                    </div>
-                  </div>
-                ))
-              )}
-            </div>
-          </div>
-        </div>
-
-        {/* Enhanced Legend - Collapsible */}
-        <Collapsible className="mt-4 border-t pt-4">
-          <CollapsibleTrigger className="group flex w-full items-center justify-between">
-            <div className="flex items-center gap-2">
-              <h5 className="text-sm font-medium">Status Legend</h5>
-              <ChevronDown className="h-4 w-4 text-muted-foreground transition-transform group-data-[state=open]:rotate-180" />
-            </div>
-            <span className="text-xs text-muted-foreground">
-              Hover over tasks for detailed timeline
-            </span>
-          </CollapsibleTrigger>
-
-          <CollapsibleContent className="mt-3">
-            <div className="flex flex-wrap items-center gap-6 text-xs">
-              <div className="flex items-center gap-2">
-                <div className="h-3 w-3 rounded bg-gray-400"></div>
-                <span>Not Started</span>
-              </div>
-              <div className="flex items-center gap-2">
-                <div className="h-3 w-3 rounded bg-blue-500"></div>
-                <span>Active</span>
-              </div>
-              <div className="flex items-center gap-2">
-                <div className="h-3 w-3 rounded bg-green-500"></div>
-                <span>Completed</span>
-              </div>
-              <div className="flex items-center gap-2">
-                <div className="h-1 w-3 rounded bg-yellow-400"></div>
-                <span>Status Change</span>
-              </div>
-              <div className="flex items-center gap-2">
-                <div className="h-1 w-3 rounded bg-gradient-to-r from-gray-300 via-blue-400 to-green-400"></div>
-                <span>Lifecycle Phases</span>
-              </div>
-              <div className="flex items-center gap-2">
-                <div className="h-1 w-3 rounded bg-red-500"></div>
-                <span>Overdue</span>
-              </div>
-            </div>
-          </CollapsibleContent>
-        </Collapsible>
-      </Card>
-
-      {/* Beautiful Click Card */}
-      {clickCardVisible && clickedTask && (
-        <>
-          {/* Subtle backdrop */}
-          <div
-            className="fixed inset-0 z-[9998] bg-black/5 backdrop-blur-[1px]"
-            onClick={handleCloseClick}
-          />
-
-          <div
-            ref={clickCardRef}
-            className="fixed z-[9999] max-w-sm rounded-lg border border-gray-200 bg-white shadow-xl duration-200 animate-in slide-in-from-bottom-2 dark:border-gray-700 dark:bg-gray-900"
-            style={{
-              left: Math.max(
-                10,
-                Math.min(clickCardPosition.x, window.innerWidth - 350)
-              ),
-              top: Math.max(
-                10,
-                Math.min(clickCardPosition.y, window.innerHeight - 200)
-              ),
-            }}
-          >
-            {/* Compact Header */}
-            <div className="rounded-t-lg border-b border-gray-100 bg-gradient-to-r from-blue-50 to-indigo-50 p-3 dark:border-gray-800 dark:from-blue-900/20 dark:to-indigo-900/20">
-              <div className="flex items-center justify-between">
-                <div className="min-w-0 flex-1">
-                  <h4 className="truncate text-sm font-semibold text-gray-800 dark:text-gray-100">
-                    {clickedTask.name}
-                  </h4>
-                  <div className="mt-1 flex items-center gap-1">
-                    <Badge variant="outline" className="px-1 py-0 text-xs">
-                      {clickedTask.boardName || 'Unknown Board'}
-                    </Badge>
-                    <span className="text-xs text-muted-foreground">•</span>
-                    <Badge variant="outline" className="px-1 py-0 text-xs">
-                      {clickedTask.listName || 'Unknown List'}
-                    </Badge>
-                  </div>
-                </div>
-                <div className="flex items-center gap-2">
-                  <Badge
-                    variant="outline"
-                    className={cn(
-                      'text-xs font-medium',
-                      clickedTask.status === 'done' &&
-                        'border-green-200 bg-green-50 text-green-700 dark:border-green-800 dark:bg-green-900/20 dark:text-green-400',
-                      clickedTask.status === 'closed' &&
-                        'border-purple-200 bg-purple-50 text-purple-700 dark:border-purple-800 dark:bg-purple-900/20 dark:text-purple-400',
-                      clickedTask.status === 'active' &&
-                        'border-blue-200 bg-blue-50 text-blue-700 dark:border-blue-800 dark:bg-blue-900/20 dark:text-blue-400',
-                      clickedTask.status === 'not_started' &&
-                        'border-gray-200 bg-gray-50 text-gray-700 dark:border-gray-800 dark:bg-gray-900/20 dark:text-gray-400'
-                    )}
-                  >
-                    {clickedTask.status === 'done' ? (
-                      <span className="flex items-center gap-1">
-                        <span className="h-1.5 w-1.5 rounded-full bg-green-500"></span>
-                        Done
-                      </span>
-                    ) : clickedTask.status === 'closed' ? (
-                      <span className="flex items-center gap-1">
-                        <span className="h-1.5 w-1.5 rounded-full bg-purple-500"></span>
-                        Closed
-                      </span>
-                    ) : clickedTask.status === 'active' ? (
-                      <span className="flex items-center gap-1">
-                        <span className="h-1.5 w-1.5 animate-pulse rounded-full bg-blue-500"></span>
-                        Active
-                      </span>
-                    ) : (
-                      <span className="flex items-center gap-1">
-                        <span className="h-1.5 w-1.5 rounded-full bg-gray-400"></span>
-                        Pending
-                      </span>
-                    )}
-                  </Badge>
-                  <button
-                    onClick={handleCloseClick}
-                    className="text-gray-400 transition-colors hover:text-gray-600 dark:hover:text-gray-300"
-                  >
-                    <X className="h-4 w-4" />
-                  </button>
-                </div>
-              </div>
-            </div>
-
-            {/* Compact Content */}
-            <div className="space-y-2 p-3">
-              {clickedTask.description && (
-                <p className="line-clamp-2 text-xs text-gray-600 dark:text-gray-400">
-                  {clickedTask.description}
-                </p>
-              )}
-
-              {/* Compact Info Grid */}
-              <div className="grid grid-cols-2 gap-2 text-xs">
-                <div className="space-y-1">
-                  <div className="flex items-center gap-1">
-                    <Calendar className="h-3 w-3 text-green-500" />
-                    <span className="text-muted-foreground">Created:</span>
-                  </div>
-                  <div className="pl-4 font-medium">
-                    {clickedTask.createdDate
-                      ? clickedTask.createdDate.toLocaleDateString()
-                      : 'N/A'}
-                  </div>
-
-                  {clickedTask.end_date && (
-                    <>
-                      <div className="flex items-center gap-1">
-                        <Clock className="h-3 w-3 text-orange-500" />
-                        <span className="text-muted-foreground">Due:</span>
-                      </div>
-                      <div className="pl-4 font-medium">
-                        {new Date(clickedTask.end_date).toLocaleDateString()}
-                      </div>
-                    </>
-                  )}
-                </div>
-
-                <div className="space-y-1">
-                  <div className="flex items-center gap-1">
-                    <Activity className="h-3 w-3 text-blue-500" />
-                    <span className="text-muted-foreground">Duration:</span>
-                  </div>
-                  <div className="pl-4 font-medium">
-                    {useMemo(() => {
-                      try {
-                        const createdDate = clickedTask.created_at ? new Date(clickedTask.created_at) : null;
-                        
-                        if (!createdDate || isNaN(createdDate.getTime())) {
-                          return 'N/A';
-                        }
-
-                        // For completed tasks, calculate actual duration
-                        if (clickedTask.status === 'done' || clickedTask.status === 'closed') {
-                          const completionDate = getTaskCompletionDate(clickedTask);
-                          if (completionDate && !isNaN(completionDate.getTime())) {
-                            const durationMs = completionDate.getTime() - createdDate.getTime();
-                            const days = Math.floor(durationMs / (1000 * 60 * 60 * 24));
-                            const hours = Math.floor(durationMs / (1000 * 60 * 60));
-                            
-                            if (days >= 1) {
-                              return `${days} day${days !== 1 ? 's' : ''}`;
-                            } else if (hours >= 1) {
-                              return `${hours} hr${hours !== 1 ? 's' : ''}`;
-                            } else {
-                              return '<1 hr';
-                            }
-                          }
-                        }
-                        
-                        // For non-completed tasks, show planned duration or time elapsed
-                        if (clickedTask.end_date) {
-                          const dueDate = new Date(clickedTask.end_date);
-                          if (!isNaN(dueDate.getTime())) {
-                            const plannedDurationMs = dueDate.getTime() - createdDate.getTime();
-                            const days = Math.floor(plannedDurationMs / (1000 * 60 * 60 * 24));
-                            const hours = Math.floor(plannedDurationMs / (1000 * 60 * 60));
-                            
-                            if (days >= 1) {
-                              return `${days} day${days !== 1 ? 's' : ''} (planned)`;
-                            } else if (hours >= 1) {
-                              return `${hours} hr${hours !== 1 ? 's' : ''} (planned)`;
-                            } else {
-                              return '<1 hr (planned)';
-                            }
-                          }
-                        }
-                        
-                        // Fallback: show time elapsed since creation
-                        const now = new Date();
-                        const elapsedMs = now.getTime() - createdDate.getTime();
-                        const days = Math.floor(elapsedMs / (1000 * 60 * 60 * 24));
-                        const hours = Math.floor(elapsedMs / (1000 * 60 * 60));
-                        
-                        if (days >= 1) {
-                          return `${days} day${days !== 1 ? 's' : ''} (elapsed)`;
-                        } else if (hours >= 1) {
-                          return `${hours} hr${hours !== 1 ? 's' : ''} (elapsed)`;
-                        } else {
-                          return '<1 hr (elapsed)';
-                        }
-                      } catch (error) {
-                        return 'N/A';
-                      }
-                    }, [clickedTask, getTaskCompletionDate])}
-                  </div>
-
-                  {clickedTask.priority && (
-                    <>
-                      <div className="flex items-center gap-1">
-                        <span className="text-muted-foreground">Priority:</span>
-                      </div>
-                      <div className="pl-4">
-                        <Badge
-                          variant={
-                            clickedTask.priority === 1
-                              ? 'destructive'
-                              : clickedTask.priority === 2
-                                ? 'default'
-                                : clickedTask.priority === 3
-                                  ? 'secondary'
-                                  : 'outline'
-                          }
-                          className="text-xs"
-                        >
-                          {clickedTask.priority === 1
-                            ? '🔥 Urgent'
-                            : clickedTask.priority === 2
-                              ? '⚡ High'
-                              : clickedTask.priority === 3
-                                ? '📋 Medium'
-                                : '📝 Low'}
-                        </Badge>
-                      </div>
-                    </>
-                  )}
-                </div>
-              </div>
-
-              {clickedTask.assignee_name && (
-                <div className="flex items-center gap-2 pt-1">
-                  <div className="flex h-5 w-5 items-center justify-center rounded-full bg-gradient-to-br from-blue-400 to-purple-500 text-xs font-medium text-white">
-                    {clickedTask.assignee_name.charAt(0).toUpperCase()}
-                  </div>
-                  <span className="text-xs font-medium">
-                    {clickedTask.assignee_name}
-                  </span>
-                </div>
-              )}
-
-              {/* Compact Progress */}
-              <div className="h-1.5 w-full rounded-full bg-gray-200 dark:bg-gray-600">
-                <div
-                  className={cn(
-                    'h-1.5 rounded-full transition-all',
-                    clickedTask.status === 'done' ||
-                      clickedTask.status === 'closed'
-                      ? 'bg-green-500'
-                      : clickedTask.status === 'active'
-                        ? 'bg-blue-500'
-                        : 'bg-gray-400 dark:bg-gray-500'
-                  )}
-                  style={{
-                    width:
-                      clickedTask.status === 'done' ||
-                      clickedTask.status === 'closed'
-                        ? '100%'
-                        : clickedTask.status === 'active'
-                          ? '60%'
-                          : '10%',
-                  }}
-                />
-              </div>
-
-              {/* Compact Overdue Warning */}
-              {clickedTask.end_date &&
-                new Date(clickedTask.end_date) < new Date() &&
-                clickedTask.status !== 'done' &&
-                clickedTask.status !== 'closed' && (
-                  <div className="flex items-center gap-1 rounded border border-red-200 bg-red-50 p-2 dark:border-red-800 dark:bg-red-900/20">
-                    <AlertTriangle className="h-3 w-3 text-red-500" />
-                    <span className="text-xs font-medium text-red-600 dark:text-red-400">
-                      Overdue by{' '}
-                      {Math.ceil(
-                        (new Date().getTime() -
-                          new Date(clickedTask.end_date).getTime()) /
-                          (1000 * 60 * 60 * 24)
-                      )}{' '}
-                      days
-                    </span>
-                  </div>
-                )}
-            </div>
-          </div>
-        </>
-      )}
-    </>
-  );
-}
 
 interface EnhancedBoardsViewProps {
   data: (TaskBoard & {
@@ -1148,206 +96,7 @@ interface TaskModalState {
   selectedBoard: string | null; // null means all boards
 }
 
-// Task Workflow & Efficiency Analytics Component
-function TaskWorkflowAnalytics({
-  allTasks,
-  selectedBoard,
-}: {
-  allTasks: any[];
-  selectedBoard: string | null;
-}) {
-  const filteredTasks = useMemo(() => {
-    if (!selectedBoard) return allTasks;
-    return allTasks.filter((task) => task.boardId === selectedBoard);
-  }, [allTasks, selectedBoard]);
 
-  const workflowAnalytics = useMemo(() => {
-    const now = new Date();
-    const oneWeekAgo = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-
-    // Task efficiency metrics
-    const completedTasks = filteredTasks.filter(
-      (task) => task.listStatus === 'done' || task.listStatus === 'closed' || task.archived
-    );
-
-    const activeTasks = filteredTasks.filter(
-      (task) => task.listStatus === 'active'
-    );
-
-    const notStartedTasks = filteredTasks.filter(
-      (task) => task.listStatus === 'not_started' || !task.listStatus
-    );
-
-    // Cycle time calculation (creation to completion)
-    const cycleTimeData = completedTasks
-      .filter((task) => task.created_at && task.updated_at)
-      .map((task) => {
-        const created = new Date(task.created_at);
-        const completed = new Date(task.updated_at);
-        return (completed.getTime() - created.getTime()) / (1000 * 60 * 60 * 24);
-      });
-
-    const avgCycleTime = cycleTimeData.length > 0 
-      ? cycleTimeData.reduce((sum, time) => sum + time, 0) / cycleTimeData.length
-      : 0;
-
-    // Task age distribution
-    const taskAges = filteredTasks
-      .filter((task) => task.created_at)
-      .map((task) => {
-        const created = new Date(task.created_at);
-        return (now.getTime() - created.getTime()) / (1000 * 60 * 60 * 24);
-      });
-
-    const newTasks = taskAges.filter((age) => age <= 3).length; // 0-3 days
-    const recentTasks = taskAges.filter((age) => age > 3 && age <= 7).length; // 4-7 days
-    const oldTasks = taskAges.filter((age) => age > 7 && age <= 30).length; // 1-4 weeks
-    const staleTasks = taskAges.filter((age) => age > 30).length; // >1 month
-
-    // Workload distribution
-    const totalTasks = filteredTasks.length;
-    const completionRate = totalTasks > 0 ? (completedTasks.length / totalTasks) * 100 : 0;
-
-    // Recent activity
-    const recentActivity = filteredTasks.filter((task) =>
-      task.updated_at && new Date(task.updated_at) >= oneWeekAgo
-    ).length;
-
-    // Priority distribution effectiveness
-    const highPriorityCompleted = completedTasks.filter((task) => task.priority === 1 || task.priority === 2).length;
-    const totalHighPriority = filteredTasks.filter((task) => task.priority === 1 || task.priority === 2).length;
-    const priorityEfficiency = totalHighPriority > 0 ? (highPriorityCompleted / totalHighPriority) * 100 : 0;
-
-    return {
-      totalTasks,
-      completedTasks: completedTasks.length,
-      activeTasks: activeTasks.length,
-      notStartedTasks: notStartedTasks.length,
-      avgCycleTime,
-      completionRate,
-      recentActivity,
-      priorityEfficiency,
-      ageDistribution: {
-        new: newTasks,
-        recent: recentTasks,
-        old: oldTasks,
-        stale: staleTasks,
-      },
-    };
-  }, [filteredTasks]);
-
-  const ageConfig = [
-    { key: 'new', label: '🆕 New (0-3d)', count: workflowAnalytics.ageDistribution.new, color: 'bg-green-500' },
-    { key: 'recent', label: '📅 Recent (4-7d)', count: workflowAnalytics.ageDistribution.recent, color: 'bg-blue-500' },
-    { key: 'old', label: '📚 Old (1-4w)', count: workflowAnalytics.ageDistribution.old, color: 'bg-yellow-500' },
-    { key: 'stale', label: '⏰ Stale (>1m)', count: workflowAnalytics.ageDistribution.stale, color: 'bg-red-500' },
-  ];
-
-  return (
-    <Card className="p-4">
-      <div className="mb-3 flex items-center justify-between">
-        <h4 className="font-medium">Workflow & Efficiency</h4>
-        <span className="text-xs text-muted-foreground">
-          {workflowAnalytics.totalTasks} total tasks
-        </span>
-      </div>
-      
-      <div className="space-y-4">
-        {/* Key Metrics */}
-        <div className="grid grid-cols-2 gap-3">
-          <div className="rounded-lg bg-gradient-to-br from-emerald-50 to-emerald-100 p-3 dark:from-emerald-900/20 dark:to-emerald-800/20">
-            <div className="text-lg font-bold text-emerald-600">
-              {workflowAnalytics.completionRate.toFixed(0)}%
-            </div>
-            <div className="text-xs text-emerald-700 dark:text-emerald-300">
-              Completion Rate
-            </div>
-          </div>
-          <div className="rounded-lg bg-gradient-to-br from-violet-50 to-violet-100 p-3 dark:from-violet-900/20 dark:to-violet-800/20">
-            <div className="text-lg font-bold text-violet-600">
-              {workflowAnalytics.avgCycleTime > 0 ? `${workflowAnalytics.avgCycleTime.toFixed(1)}d` : 'N/A'}
-            </div>
-            <div className="text-xs text-violet-700 dark:text-violet-300">
-              Avg. Cycle Time
-            </div>
-          </div>
-        </div>
-
-        {/* Status Breakdown */}
-        <div className="space-y-2">
-          <div className="text-sm font-medium">Task Status</div>
-          <div className="grid grid-cols-3 gap-2 text-xs">
-            <div className="text-center rounded-lg bg-gray-50 p-2 dark:bg-gray-800/50">
-              <div className="font-medium text-green-600">
-                {workflowAnalytics.completedTasks}
-              </div>
-              <div className="text-muted-foreground">Done</div>
-            </div>
-            <div className="text-center rounded-lg bg-gray-50 p-2 dark:bg-gray-800/50">
-              <div className="font-medium text-blue-600">
-                {workflowAnalytics.activeTasks}
-              </div>
-              <div className="text-muted-foreground">Active</div>
-            </div>
-            <div className="text-center rounded-lg bg-gray-50 p-2 dark:bg-gray-800/50">
-              <div className="font-medium text-gray-600">
-                {workflowAnalytics.notStartedTasks}
-              </div>
-              <div className="text-muted-foreground">Not Started</div>
-            </div>
-          </div>
-        </div>
-
-        {/* Task Age Distribution */}
-        <div className="space-y-2">
-          <div className="text-sm font-medium">Task Age</div>
-          <div className="space-y-1">
-            {ageConfig.filter(age => age.count > 0).map((age) => (
-              <div key={age.key} className="flex items-center justify-between">
-                <div className="flex items-center gap-2">
-                  <div className={cn('h-2 w-2 rounded', age.color)}></div>
-                  <span className="text-xs">{age.label}</span>
-                </div>
-                <span className="text-xs font-medium">{age.count}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {/* Efficiency Indicators */}
-        <div className="space-y-2">
-          <div className="text-sm font-medium">Efficiency Metrics</div>
-          <div className="grid grid-cols-2 gap-2 text-xs">
-            <div className="rounded-lg bg-orange-50 p-2 dark:bg-orange-900/20">
-              <div className="font-medium text-orange-600">
-                {workflowAnalytics.recentActivity}
-              </div>
-              <div className="text-orange-700 dark:text-orange-300">Recent Updates</div>
-            </div>
-            <div className="rounded-lg bg-indigo-50 p-2 dark:bg-indigo-900/20">
-              <div className="font-medium text-indigo-600">
-                {workflowAnalytics.priorityEfficiency.toFixed(0)}%
-              </div>
-              <div className="text-indigo-700 dark:text-indigo-300">Priority Focus</div>
-            </div>
-          </div>
-        </div>
-
-        {/* Stale Task Warning */}
-        {workflowAnalytics.ageDistribution.stale > 0 && (
-          <div className="rounded-lg border border-red-200 bg-red-50 p-2 dark:border-red-800 dark:bg-red-900/20">
-            <div className="flex items-center gap-2">
-              <Clock className="h-3 w-3 text-red-600" />
-              <span className="text-xs font-medium text-red-700 dark:text-red-300">
-                {workflowAnalytics.ageDistribution.stale} stale tasks need attention
-              </span>
-            </div>
-          </div>
-        )}
-      </div>
-    </Card>
-  );
-}
 
 export function EnhancedBoardsView({ data, count }: EnhancedBoardsViewProps) {
   const [selectedBoard, setSelectedBoard] = useState<(typeof data)[0] | null>(null);
@@ -2963,534 +1712,10 @@ export function EnhancedBoardsView({ data, count }: EnhancedBoardsViewProps) {
   );
 }
 
-// Task Group Component for collapsible sections
-interface TaskGroupProps {
-  title: string;
-  icon: React.ReactNode;
-  tasks: Array<{
-    id: string;
-    name: string;
-    description?: string;
-    priority?: number | null;
-    end_date?: string | null;
-    boardName: string;
-    listName: string;
-    boardHref?: string;
-    archived?: boolean;
-    listStatus?: string;
-  }>;
-  count: number;
-  onTaskClick: (task: any) => void;
-}
-
-function TaskGroup({ title, icon, tasks, count, onTaskClick }: TaskGroupProps) {
-  const [isOpen, setIsOpen] = useState(true);
-  const [expandedTasks, setExpandedTasks] = useState<Set<string>>(new Set());
-
-  if (count === 0) return null;
-
-  const toggleTaskExpansion = (taskId: string, e: React.MouseEvent) => {
-    e.stopPropagation();
-    setExpandedTasks(prev => {
-      const newSet = new Set(prev);
-      if (newSet.has(taskId)) {
-        newSet.delete(taskId);
-      } else {
-        newSet.add(taskId);
-      }
-      return newSet;
-    });
-  };
-
-  return (
-    <Collapsible open={isOpen} onOpenChange={setIsOpen}>
-      <CollapsibleTrigger asChild>
-        <Button
-          variant="ghost"
-          className="h-auto w-full justify-between p-3 hover:bg-muted/50"
-        >
-          <div className="flex items-center gap-3">
-            {icon}
-            <span className="font-medium">{title}</span>
-            <Badge variant="secondary" className="ml-2">
-              {count}
-            </Badge>
-          </div>
-          {isOpen ? (
-            <ChevronDown className="h-4 w-4" />
-          ) : (
-            <ChevronRight className="h-4 w-4" />
-          )}
-        </Button>
-      </CollapsibleTrigger>
-
-      <CollapsibleContent className="space-y-2 pt-2">
-        {tasks.map((task) => {
-          const isExpanded = expandedTasks.has(task.id);
-          const hasLongContent = task.name.length > 60 || (task.description && task.description.length > 100);
-          
-          return (
-            <Card
-              key={task.id}
-              className="group cursor-pointer p-4 transition-all duration-200 hover:bg-muted/50 hover:shadow-sm"
-              onClick={() => onTaskClick(task)}
-            >
-              <div className="flex items-start gap-4">
-                <div className="min-w-0 flex-1 space-y-3">
-                  {/* Task Title and Priority */}
-                  <div className="flex items-start gap-2">
-                    <div className="min-w-0 flex-1">
-                      <h4 
-                        className={cn(
-                          "text-sm font-medium leading-relaxed transition-all duration-200",
-                          isExpanded ? "line-clamp-none" : "line-clamp-2"
-                        )}
-                        title={task.name}
-                      >
-                        {task.name}
-                      </h4>
-                    </div>
-                    
-                    <div className="flex items-center gap-2 flex-shrink-0">
-                      {task.priority === 1 && (
-                        <Badge variant="destructive" className="text-xs whitespace-nowrap">
-                          🔥 Urgent
-                        </Badge>
-                      )}
-                      {task.priority === 2 && (
-                        <Badge variant="secondary" className="text-xs whitespace-nowrap">
-                          ⚡ High
-                        </Badge>
-                      )}
-                      {task.priority === 3 && (
-                        <Badge variant="outline" className="text-xs whitespace-nowrap">
-                          📋 Medium
-                        </Badge>
-                      )}
-                      {task.priority === 4 && (
-                        <Badge variant="outline" className="text-xs whitespace-nowrap">
-                          📝 Low
-                        </Badge>
-                      )}
-                      
-                      {/* Expand/Collapse button for long content */}
-                      {hasLongContent && (
-                        <Button
-                          variant="ghost"
-                          size="sm"
-                          className="h-6 w-6 p-0 text-muted-foreground hover:text-foreground"
-                          onClick={(e) => toggleTaskExpansion(task.id, e)}
-                        >
-                          {isExpanded ? (
-                            <ChevronDown className="h-3 w-3" />
-                          ) : (
-                            <ChevronRight className="h-3 w-3" />
-                          )}
-                        </Button>
-                      )}
-                    </div>
-                  </div>
-
-                  {/* Task Description */}
-                  {task.description && (
-                    <div className="text-xs text-muted-foreground">
-                      <p 
-                        className={cn(
-                          "leading-relaxed transition-all duration-200",
-                          isExpanded ? "line-clamp-none" : "line-clamp-2"
-                        )}
-                        title={task.description}
-                      >
-                        {task.description}
-                      </p>
-                    </div>
-                  )}
-
-                  {/* Task Metadata */}
-                  <div className="flex flex-wrap items-center gap-3 text-xs text-muted-foreground">
-                    <span className="flex items-center gap-1 min-w-0">
-                      <LayoutGrid className="h-3 w-3 flex-shrink-0" />
-                      <span className="truncate" title={task.boardName}>
-                        {task.boardName}
-                      </span>
-                    </span>
-                    
-                    <span className="flex items-center gap-1 min-w-0">
-                      <LayoutList className="h-3 w-3 flex-shrink-0" />
-                      <span className="truncate" title={task.listName}>
-                        {task.listName}
-                      </span>
-                    </span>
-                    
-                    {task.end_date && (
-                      <span className="flex items-center gap-1 whitespace-nowrap">
-                        <Calendar className="h-3 w-3 flex-shrink-0" />
-                        <span>
-                          {new Date(task.end_date).toLocaleDateString()}
-                        </span>
-                      </span>
-                    )}
-                  </div>
-
-                  {/* Overdue Warning */}
-                  {task.end_date && 
-                   new Date(task.end_date) < new Date() && 
-                   !task.archived && 
-                   task.listStatus !== 'done' && 
-                   task.listStatus !== 'closed' && (
-                    <div className="flex items-center gap-1 rounded-md bg-red-50 px-2 py-1 dark:bg-red-900/20">
-                      <AlertTriangle className="h-3 w-3 text-red-500" />
-                      <span className="text-xs font-medium text-red-600 dark:text-red-400">
-                        Overdue by{' '}
-                        {typeof task.end_date === 'string' ? calculateOverdueDays(task.end_date) : 0}{' '}
-                        days
-                      </span>
-                    </div>
-                  )}
-                </div>
-
-                {/* Action Button */}
-                <div className="flex-shrink-0">
-                  <Button
-                    variant="ghost"
-                    size="sm"
-                    className="h-8 w-8 p-0 opacity-60 transition-opacity group-hover:opacity-100"
-                    onClick={(e) => {
-                      e.stopPropagation();
-                      onTaskClick(task);
-                    }}
-                  >
-                    <ExternalLink className="h-4 w-4" />
-                  </Button>
-                </div>
-              </div>
-            </Card>
-          );
-        })}
-      </CollapsibleContent>
-    </Collapsible>
-  );
-}
-
-// Status Distribution Component
-function StatusDistribution({
-  allTasks,
-  selectedBoard,
-}: {
-  allTasks: any[];
-  selectedBoard: string | null;
-}) {
-  const filteredTasks = useMemo(() => {
-    if (!selectedBoard) return allTasks;
-    return allTasks.filter((task) => task.boardId === selectedBoard);
-  }, [allTasks, selectedBoard]);
-
-  const statusCounts = useMemo(() => {
-    const counts = {
-      not_started: 0,
-      active: 0,
-      done: 0,
-      closed: 0,
-    };
-
-    filteredTasks.forEach((task) => {
-      const status = task.listStatus || 'not_started';
-      if (status === 'done' || task.archived) {
-        counts.done += 1;
-      } else if (status === 'closed') {
-        counts.closed += 1;
-      } else if (status === 'active') {
-        counts.active += 1;
-      } else {
-        counts.not_started += 1;
-      }
-    });
-
-    return counts;
-  }, [filteredTasks]);
-
-  const total = Object.values(statusCounts).reduce(
-    (sum, count) => sum + count,
-    0
-  );
-
-  const statusConfig = [
-    {
-      key: 'not_started',
-      label: 'Not Started',
-      color: 'bg-gray-400',
-      percentage: total > 0 ? (statusCounts.not_started / total) * 100 : 0,
-    },
-    {
-      key: 'active',
-      label: 'Active',
-      color: 'bg-blue-500',
-      percentage: total > 0 ? (statusCounts.active / total) * 100 : 0,
-    },
-    {
-      key: 'done',
-      label: 'Done',
-      color: 'bg-green-500',
-      percentage: total > 0 ? (statusCounts.done / total) * 100 : 0,
-    },
-    {
-      key: 'closed',
-      label: 'Closed',
-      color: 'bg-purple-500',
-      percentage: total > 0 ? (statusCounts.closed / total) * 100 : 0,
-    },
-  ];
-
-  return (
-    <Card className="p-4">
-      <div className="mb-3 flex items-center justify-between">
-        <h4 className="font-medium">Status Distribution</h4>
-        <span className="text-xs text-muted-foreground">{total} tasks</span>
-      </div>
-      <div className="space-y-3">
-        {statusConfig.map((status) => (
-          <div key={status.key} className="space-y-1">
-            <div className="flex items-center justify-between">
-              <div className="flex items-center gap-2">
-                <div className={cn('h-3 w-3 rounded', status.color)}></div>
-                <span className="text-sm">{status.label}</span>
-              </div>
-              <div className="flex items-center gap-2">
-                <span className="text-sm font-medium">
-                  {statusCounts[status.key as keyof typeof statusCounts]}
-                </span>
-                <span className="text-xs text-muted-foreground">
-                  ({status.percentage.toFixed(0)}%)
-                </span>
-              </div>
-            </div>
-            {/* Progress bar */}
-            <div className="h-1.5 w-full rounded-full bg-gray-200 dark:bg-gray-700">
-              <div
-                className={cn(
-                  'h-1.5 rounded-full transition-all',
-                  status.color
-                )}
-                style={{ width: `${status.percentage}%` }}
-              />
-            </div>
-          </div>
-        ))}
-        {total === 0 && (
-          <div className="py-4 text-center text-sm text-muted-foreground">
-            No tasks found
-          </div>
-        )}
-      </div>
-    </Card>
-  );
-}
 
 
 
-// Task Creation & Flow Analytics Component
-function TaskCreationAnalytics({
-  allTasks,
-  selectedBoard,
-}: {
-  allTasks: any[];
-  selectedBoard: string | null;
-}) {
-  const filteredTasks = useMemo(() => {
-    if (!selectedBoard) return allTasks;
-    return allTasks.filter((task) => task.boardId === selectedBoard);
-  }, [allTasks, selectedBoard]);
 
-  const taskAnalytics = useMemo(() => {
-    const now = new Date();
-    const oneWeekAgo = new Date(now.getTime() - 7 * 24 * 60 * 60 * 1000);
-    const twoWeeksAgo = new Date(now.getTime() - 14 * 24 * 60 * 60 * 1000);
-    const oneMonthAgo = new Date(now.getTime() - 30 * 24 * 60 * 60 * 1000);
 
-    // Task creation trends
-    const thisWeekCreated = filteredTasks.filter((task) => {
-      return task.created_at && new Date(task.created_at) >= oneWeekAgo;
-    }).length;
 
-    const lastWeekCreated = filteredTasks.filter((task) => {
-      return (
-        task.created_at &&
-        new Date(task.created_at) >= twoWeeksAgo &&
-        new Date(task.created_at) < oneWeekAgo
-      );
-    }).length;
 
-    const thisMonthCreated = filteredTasks.filter((task) => {
-      return task.created_at && new Date(task.created_at) >= oneMonthAgo;
-    }).length;
-
-    // Task distribution by priority
-    const priorityDistribution = {
-      urgent: filteredTasks.filter((task) => task.priority === 1).length,
-      high: filteredTasks.filter((task) => task.priority === 2).length,
-      medium: filteredTasks.filter((task) => task.priority === 3).length,
-      low: filteredTasks.filter((task) => task.priority === 4).length,
-      unset: filteredTasks.filter((task) => !task.priority).length,
-    };
-
-    // Task type distribution by board lists
-    const typeDistribution: { [key: string]: number } = {};
-    filteredTasks.forEach((task) => {
-      const listName = task.listName || 'Unknown';
-      typeDistribution[listName] = (typeDistribution[listName] || 0) + 1;
-    });
-
-    // Average time to start (created_at to first status change)
-    const timeToStartData = filteredTasks
-      .filter((task) => 
-        task.created_at && 
-        task.updated_at && 
-        task.listStatus !== 'not_started' &&
-        new Date(task.updated_at).getTime() > new Date(task.created_at).getTime()
-      )
-      .map((task) => {
-        const created = new Date(task.created_at);
-        const updated = new Date(task.updated_at);
-        return (updated.getTime() - created.getTime()) / (1000 * 60 * 60 * 24); // days
-      });
-
-    const avgTimeToStart = timeToStartData.length > 0 
-      ? timeToStartData.reduce((sum, time) => sum + time, 0) / timeToStartData.length
-      : 0;
-
-    // Task creation rate trend
-    const creationTrend = lastWeekCreated > 0 
-      ? ((thisWeekCreated - lastWeekCreated) / lastWeekCreated) * 100
-      : thisWeekCreated > 0 ? 100 : 0;
-
-    // Backlog growth (not started tasks older than 1 week)
-    const backlogTasks = filteredTasks.filter((task) => 
-      task.listStatus === 'not_started' &&
-      task.created_at &&
-      new Date(task.created_at) < oneWeekAgo
-    ).length;
-
-    return {
-      thisWeekCreated,
-      lastWeekCreated,
-      thisMonthCreated,
-      creationTrend,
-      priorityDistribution,
-      typeDistribution,
-      avgTimeToStart,
-      backlogTasks,
-      totalTasks: filteredTasks.length,
-    };
-  }, [filteredTasks]);
-
-  const priorityConfig = [
-    { key: 'urgent', label: '🔥 Urgent', color: 'bg-red-500', count: taskAnalytics.priorityDistribution.urgent },
-    { key: 'high', label: '⚡ High', color: 'bg-orange-500', count: taskAnalytics.priorityDistribution.high },
-    { key: 'medium', label: '📋 Medium', color: 'bg-yellow-500', count: taskAnalytics.priorityDistribution.medium },
-    { key: 'low', label: '📝 Low', color: 'bg-green-500', count: taskAnalytics.priorityDistribution.low },
-    { key: 'unset', label: '❓ Unset', color: 'bg-gray-400', count: taskAnalytics.priorityDistribution.unset },
-  ];
-
-  const topTaskTypes = Object.entries(taskAnalytics.typeDistribution)
-    .sort(([,a], [,b]) => b - a)
-    .slice(0, 4);
-
-  return (
-    <Card className="p-4">
-      <div className="mb-3 flex items-center justify-between">
-        <h4 className="font-medium">Task Creation & Flow</h4>
-        <span className="text-xs text-muted-foreground">
-          {taskAnalytics.totalTasks} total tasks
-        </span>
-      </div>
-      
-      <div className="space-y-4">
-        {/* Creation Trend */}
-        <div className="rounded-lg bg-blue-50 p-3 dark:bg-blue-900/20">
-          <div className="flex items-center justify-between">
-            <div>
-              <div className="text-lg font-bold text-blue-600">
-                {taskAnalytics.thisWeekCreated}
-              </div>
-              <div className="text-xs text-blue-700 dark:text-blue-300">
-                Tasks created this week
-              </div>
-            </div>
-            <div className="text-right">
-              <div className={cn(
-                "text-sm font-medium",
-                taskAnalytics.creationTrend > 0 ? "text-green-600" : 
-                taskAnalytics.creationTrend < 0 ? "text-red-600" : "text-gray-600"
-              )}>
-                {taskAnalytics.creationTrend > 0 ? '↗' : taskAnalytics.creationTrend < 0 ? '↘' : '→'} 
-                {Math.abs(taskAnalytics.creationTrend).toFixed(0)}%
-              </div>
-              <div className="text-xs text-muted-foreground">
-                vs last week ({taskAnalytics.lastWeekCreated})
-              </div>
-            </div>
-          </div>
-        </div>
-
-        {/* Quick Stats */}
-        <div className="grid grid-cols-2 gap-3 text-xs">
-          <div className="text-center rounded-lg bg-gray-50 p-2 dark:bg-gray-800/50">
-            <div className="font-medium text-purple-600">
-              {taskAnalytics.thisMonthCreated}
-            </div>
-            <div className="text-muted-foreground">This month</div>
-          </div>
-          <div className="text-center rounded-lg bg-gray-50 p-2 dark:bg-gray-800/50">
-            <div className="font-medium text-orange-600">
-              {taskAnalytics.avgTimeToStart > 0 ? `${taskAnalytics.avgTimeToStart.toFixed(1)}d` : 'N/A'}
-            </div>
-            <div className="text-muted-foreground">Avg. time to start</div>
-          </div>
-        </div>
-
-        {/* Priority Distribution */}
-        <div className="space-y-2">
-          <div className="text-sm font-medium">Priority Distribution</div>
-          <div className="space-y-1">
-            {priorityConfig.filter(p => p.count > 0).map((priority) => (
-              <div key={priority.key} className="flex items-center justify-between">
-                <div className="flex items-center gap-2">
-                  <div className={cn('h-2 w-2 rounded', priority.color)}></div>
-                  <span className="text-xs">{priority.label}</span>
-                </div>
-                <span className="text-xs font-medium">{priority.count}</span>
-              </div>
-            ))}
-          </div>
-        </div>
-
-        {/* Task Types */}
-        {topTaskTypes.length > 0 && (
-          <div className="space-y-2">
-            <div className="text-sm font-medium">Top Task Types</div>
-            <div className="space-y-1">
-              {topTaskTypes.map(([type, count]) => (
-                <div key={type} className="flex items-center justify-between">
-                  <span className="text-xs truncate flex-1">{type}</span>
-                  <span className="text-xs font-medium ml-2">{count}</span>
-                </div>
-              ))}
-            </div>
-          </div>
-        )}
-
-        {/* Backlog Alert */}
-        {taskAnalytics.backlogTasks > 0 && (
-          <div className="rounded-lg border border-yellow-200 bg-yellow-50 p-2 dark:border-yellow-800 dark:bg-yellow-900/20">
-            <div className="flex items-center gap-2">
-              <AlertTriangle className="h-3 w-3 text-yellow-600" />
-                             <span className="text-xs font-medium text-yellow-700 dark:text-yellow-300">
-                 {taskAnalytics.backlogTasks} tasks in backlog (&gt;1 week old)
-               </span>
-            </div>
-          </div>
-        )}
-      </div>
-    </Card>
-  );
-}
